Move geometry to QPolygonF list function to common static method
