+++ python/core/auto_generated/symbology/qgssymbollayerutils.sip.in
@@ -796,6 +796,15 @@ Determines an SVG symbol's name from its ``path``.
 .. seealso:: :py:func:`svgSymbolNameToPath`
 %End
 
+    static QList< QList< QPolygonF > > toQPolygonF( const QgsGeometry &geometry, Qgis::SymbolType type );
+%Docstring
+Converts a ``geometry`` to a set of QPolygonF objects representing
+how the geometry should be drawn for a symbol of the given ``type``,
+as a list of geometry parts and rings.
+
+.. versionadded:: 3.40
+%End
+
     static QPointF polygonCentroid( const QPolygonF &points );
 %Docstring
 Calculate the centroid point of a QPolygonF
