+++ src/core/symbology/qgssymbollayerutils.h
@@ -743,6 +743,15 @@ class CORE_EXPORT QgsSymbolLayerUtils
      */
     static QString svgSymbolPathToName( const QString &path, const QgsPathResolver &pathResolver );
 
+    /**
+     * Converts a \a geometry to a set of QPolygonF objects representing
+     * how the geometry should be drawn for a symbol of the given \a type,
+     * as a list of geometry parts and rings.
+     *
+     * \since QGIS 3.40
+     */
+    static QList< QList< QPolygonF > > toQPolygonF( const QgsGeometry &geometry, Qgis::SymbolType type );
+
     //! Calculate the centroid point of a QPolygonF
     static QPointF polygonCentroid( const QPolygonF &points );
